+++ x/perp/keeper/clearing_house_integration_test.go
@@ -278,7 +278,7 @@ func TestOpenPositionError(t *testing.T) {
 			margin:          sdk.NewInt(1000),
 			leverage:        sdk.NewDec(10),
 			baseLimit:       sdk.NewDec(10_000),
-			expectedErr:     vpooltypes.ErrAssetOverUserLimit,
+			expectedErr:     vpooltypes.ErrAssetFailsUserLimit,
 		},
 		{
 			name:            "new short position not under base limit",
@@ -288,7 +288,7 @@ func TestOpenPositionError(t *testing.T) {
 			margin:          sdk.NewInt(1000),
 			leverage:        sdk.NewDec(10),
 			baseLimit:       sdk.NewDec(10_000),
-			expectedErr:     vpooltypes.ErrAssetOverUserLimit,
+			expectedErr:     vpooltypes.ErrAssetFailsUserLimit,
 		},
 	}