+++ x/vpool/keeper/keeper.go
@@ -39,6 +39,7 @@ args:
   - dir: either add or remove from pool
   - baseAssetAmount: the amount of quote asset being traded
   - quoteAmountLimit: a limiter to ensure the trader doesn't get screwed by slippage
+  - skipFluctuationLimitCheck: whether or not to skip the fluctuation limit check
 
 ret:
   - quoteAssetAmount: the amount of quote asset swapped
@@ -50,6 +51,7 @@ func (k Keeper) SwapBaseForQuote(
 	dir types.Direction,
 	baseAssetAmount sdk.Dec,
 	quoteAmountLimit sdk.Dec,
+	skipFluctuationLimitCheck bool,
 ) (quoteAssetAmount sdk.Dec, err error) {
 	if !k.ExistsPool(ctx, pair) {
 		return sdk.Dec{}, types.ErrPairNotSupported
@@ -64,8 +66,7 @@ func (k Keeper) SwapBaseForQuote(
 		return sdk.Dec{}, err
 	}
 
-	if dir == types.Direction_REMOVE_FROM_POOL &&
-		!pool.HasEnoughBaseReserve(baseAssetAmount) {
+	if dir == types.Direction_REMOVE_FROM_POOL && !pool.HasEnoughBaseReserve(baseAssetAmount) {
 		return sdk.Dec{}, types.ErrOverTradingLimit
 	}
 
@@ -77,14 +78,14 @@ func (k Keeper) SwapBaseForQuote(
 	if !quoteAmountLimit.IsZero() {
 		// if going long and the base amount retrieved from the pool is less than the limit
 		if dir == types.Direction_ADD_TO_POOL && quoteAssetAmount.LT(quoteAmountLimit) {
-			return sdk.Dec{}, fmt.Errorf(
+			return sdk.Dec{}, types.ErrAssetFailsUserLimit.Wrapf(
 				"quote amount (%s) is less than selected limit (%s)",
 				quoteAssetAmount.String(),
 				quoteAmountLimit.String(),
 			)
 			// if going short and the base amount retrieved from the pool is greater than the limit
 		} else if dir == types.Direction_REMOVE_FROM_POOL && quoteAssetAmount.GT(quoteAmountLimit) {
-			return sdk.Dec{}, fmt.Errorf(
+			return sdk.Dec{}, types.ErrAssetFailsUserLimit.Wrapf(
 				"quote amount (%s) is greater than selected limit (%s)",
 				quoteAssetAmount.String(),
 				quoteAmountLimit.String(),
@@ -100,7 +101,7 @@ func (k Keeper) SwapBaseForQuote(
 		pool.IncreaseQuoteAssetReserve(quoteAssetAmount)
 	}
 
-	if err = k.savePoolAndSnapshot(ctx, pool, false /*skipFluctuationCheck*/); err != nil {
+	if err = k.updatePool(ctx, pool, skipFluctuationLimitCheck); err != nil {
 		return sdk.Dec{}, fmt.Errorf("error updating reserve: %w", err)
 	}
 
@@ -135,6 +136,7 @@ args:
   - dir: either add or remove from pool
   - quoteAssetAmount: the amount of quote asset being traded
   - baseAmountLimit: a limiter to ensure the trader doesn't get screwed by slippage
+  - skipFluctuationLimitCheck: whether or not to skip the fluctuation limit check
 
 ret:
   - baseAssetAmount: the amount of base asset swapped
@@ -146,6 +148,7 @@ func (k Keeper) SwapQuoteForBase(
 	dir types.Direction,
 	quoteAssetAmount sdk.Dec,
 	baseAmountLimit sdk.Dec,
+	skipFluctuationLimitCheck bool,
 ) (baseAssetAmount sdk.Dec, err error) {
 	if !k.ExistsPool(ctx, pair) {
 		return sdk.Dec{}, types.ErrPairNotSupported
@@ -173,14 +176,14 @@ func (k Keeper) SwapQuoteForBase(
 	if !baseAmountLimit.IsZero() {
 		// if going long and the base amount retrieved from the pool is less than the limit
 		if dir == types.Direction_ADD_TO_POOL && baseAssetAmount.LT(baseAmountLimit) {
-			return sdk.Dec{}, types.ErrAssetOverUserLimit.Wrapf(
+			return sdk.Dec{}, types.ErrAssetFailsUserLimit.Wrapf(
 				"base amount (%s) is less than selected limit (%s)",
 				baseAssetAmount.String(),
 				baseAmountLimit.String(),
 			)
 			// if going short and the base amount retrieved from the pool is greater than the limit
 		} else if dir == types.Direction_REMOVE_FROM_POOL && baseAssetAmount.GT(baseAmountLimit) {
-			return sdk.Dec{}, types.ErrAssetOverUserLimit.Wrapf(
+			return sdk.Dec{}, types.ErrAssetFailsUserLimit.Wrapf(
 				"base amount (%s) is greater than selected limit (%s)",
 				baseAssetAmount.String(),
 				baseAmountLimit.String(),
@@ -196,7 +199,7 @@ func (k Keeper) SwapQuoteForBase(
 		pool.DecreaseQuoteAssetReserve(quoteAssetAmount)
 	}
 
-	if err = k.savePoolAndSnapshot(ctx, pool, false /*skipFluctuationCheck*/); err != nil {
+	if err = k.updatePool(ctx, pool, skipFluctuationLimitCheck); err != nil {
 		return sdk.Dec{}, fmt.Errorf("error updating reserve: %w", err)
 	}
 
@@ -219,29 +222,60 @@ func (k Keeper) SwapQuoteForBase(
 	})
 }
 
+/**
+Check's that a pool that we're about to save to state does not violate the fluctuation limit.
+Always tries to check against a snapshot from a previous block. If one doesn't exist, then it just uses the current snapshot.
+This should run prior to updating the snapshot, otherwise it will compare the currently updated vpool to itself.
+
+args:
+  - ctx: the cosmos-sdk context
+  - pool: the updated vpool
+
+ret:
+  - err: error if any
+*/
 func (k Keeper) checkFluctuationLimitRatio(ctx sdk.Context, pool *types.Pool) error {
-	if pool.FluctuationLimitRatio.GT(sdk.ZeroDec()) {
-		latestSnapshot, counter, err := k.getLatestReserveSnapshot(ctx, pool.Pair)
+	if pool.FluctuationLimitRatio.IsZero() {
+		// early return to avoid expensive state operations
+		return nil
+	}
+
+	latestSnapshot, counter, err := k.getLatestReserveSnapshot(ctx, pool.Pair)
+	if err != nil {
+		return fmt.Errorf("error getting last snapshot number for pair %s", pool.Pair)
+	}
+
+	if latestSnapshot.BlockNumber == ctx.BlockHeight() && counter > 0 {
+		latestSnapshot, err = k.getSnapshot(ctx, pool.Pair, counter-1)
 		if err != nil {
-			return fmt.Errorf("error getting last snapshot number for pair %s", pool.Pair)
+			return fmt.Errorf("error getting snapshot number %d from pair %s", counter-1, pool.Pair)
 		}
+	}
 
-		if latestSnapshot.BlockNumber == ctx.BlockHeight() && counter > 0 {
-			latestSnapshot, err = k.getSnapshot(ctx, pool.Pair, counter-1)
-			if err != nil {
-				return fmt.Errorf("error getting snapshot number %d from pair %s", counter-1, pool.Pair)
-			}
-		}
-
-		if isOverFluctuationLimit(pool, latestSnapshot) {
-			return types.ErrOverFluctuationLimit
-		}
+	if isOverFluctuationLimit(pool, latestSnapshot) {
+		return types.ErrOverFluctuationLimit
 	}
 
 	return nil
 }
 
+/**
+isOverFluctuationLimit compares the updated pool's reserves with the given reserve snapshot, and errors if the fluctuation is above the bounds.
+
+If the fluctuation limit ratio is zero, then the fluctuation limit check is skipped.
+
+args:
+  - pool: the updated vpool
+  - snapshot: the snapshot to compare against
+
+ret:
+  - bool: true if the fluctuation limit is violated. false otherwise
+*/
 func isOverFluctuationLimit(pool *types.Pool, snapshot types.ReserveSnapshot) bool {
+	if pool.FluctuationLimitRatio.IsZero() {
+		return false
+	}
+
 	price := pool.QuoteAssetReserve.Quo(pool.BaseAssetReserve)
 
 	lastPrice := snapshot.QuoteAssetReserve.Quo(snapshot.BaseAssetReserve)
@@ -255,7 +289,18 @@ func isOverFluctuationLimit(pool *types.Pool, snapshot types.ReserveSnapshot) bo
 	return false
 }
 
-func (k Keeper) IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) (isIt bool) {
+/**
+IsOverSpreadLimit compares the current spot price of the vpool (given by pair) to the underlying's index price (given by an oracle).
+It panics if you provide it with a pair that doesn't exist in the state.
+
+args:
+  - ctx: the cosmos-sdk context
+  - pair: the asset pair
+
+ret:
+  - bool: whether or not the price has deviated from the oracle price beyond a spread ratio
+*/
+func (k Keeper) IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) bool {
 	spotPrice, err := k.GetSpotPrice(ctx, pair)
 	if err != nil {
 		panic(err)
@@ -270,5 +315,6 @@ func (k Keeper) IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) (isIt 
 	if err != nil {
 		panic(err)
 	}
+
 	return spotPrice.Sub(oraclePrice).Quo(oraclePrice).Abs().GTE(pool.MaxOracleSpreadRatio)
 }